Change js file into modules
