+++ index.html
@@ -5,9 +5,8 @@
   <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="css/layout.css">
-  <script src="scripts/main.js" defer></script>
-  <script src="scripts/scroll.js" defer></script>
+  <link rel="stylesheet" href="css/index.css">
+  <script src="scripts/index.js" type="module"></script>
   <!-- google fonts -->
   <link rel="preconnect" href="https://fonts.googleapis.com">
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
